Controllers made ready
The controllers for this "awesome" game are ready for take off. The
player can walk or run around(I don't know which one), jump and the
players field has boundaries so he cannot cheat. The walking isn't just
2d but the 2d character can walk closer towards the camera. You basicly
can walk on z-axis and that makes it special: a touch of 3d in a 2d
game.